js/src/xpconnect/tests/mochitest/test_bug505915.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=505915
-->
<head>
  <title>Test for Bug 505915</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=505915">Mozilla Bug 505915</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript;version=1.7">

/** Test for Bug 505915 **/
window.addEventListener("message", function () { gen.next() }, false);

function go() {
    var ifr = $('ifr');
    try {
        document.documentElement.appendChild(ifr.contentWindow);
        ok(false, "weird behavior");
    } catch (e) {
        ok(/NS_ERROR_XPC_SECURITY_MANAGER_VETO/.test(e),
           "threw a security exception instead of an invalid child exception");
    }

    try {
        document.createTreeWalker(ifr.contentDocument, 0, null, false);
        ok(false, "should have thrown a security exception");
    } catch (e) {
        ok(/NS_ERROR_XPC_SECURITY_MANAGER_VETO/.test(e),
           "threw a security exception instead of an invalid child exception");
    }

    var xhr = new XMLHttpRequest();

    try {
        xhr.onreadystatechange = ifr.contentWindow;
        ok(false, "weird behavior");
    } catch (e) {
        ok(/NS_ERROR_XPC_SECURITY_MANAGER_VETO/.test(e),
           "threw a security exception instead of an invalid child exception");
    }

    // Location is always wrapped, so test it separately.

    ifr.onload = null;
    var path = "/tests/js/src/xpconnect/tests/mochitest/file_bug505915.html";

    ifr.contentWindow.location = "http://mochi.test:8888/" + path;
    yield;
    try {
        document.documentElement.appendChild(ifr.contentWindow.location);
        ok(false, "weird behavior");
    } catch (e) {
        ok(!/NS_ERROR_XPC_SECURITY_MANAGER_VETO/.test(e),
           "didn't throw a security exception");
    }
    oldLocation = ifr.contentWindow.location;
    ifr.contentWindow.location = "http://example.org/" + path;
    yield;
    try {
        document.documentElement.appendChild(oldLocation);
        ok(false, "weird behavior");
    } catch (e) {
        ok(/NS_ERROR_XPC_SECURITY_MANAGER_VETO/.test(e),
           "threw a security exception instead of an invalid child exception");
    }
    try {
        document.documentElement.appendChild(ifr.contentWindow.location);
        ok(false, "weird behavior");
    } catch (e) {
        ok(/NS_ERROR_XPC_SECURITY_MANAGER_VETO/.test(e),
           "threw a security exception instead of an invalid child exception");
    }
    oldLocation = ifr.contentWindow.location;
    ifr.contentWindow.location = "http://mochi.test:8888/" + path;
    yield;
    try {
        document.documentElement.appendChild(oldLocation);
        ok(false, "weird behavior");
    } catch (e) {
        ok(!/NS_ERROR_XPC_SECURITY_MANAGER_VETO/.test(e),
           "didn't throw a security exception");
    }

    SimpleTest.finish();
    yield;
}

SimpleTest.waitForExplicitFinish();

</script>
</pre>

<iframe id="ifr" onload="gen = go(); gen.next();" src="http://example.org/"></iframe>

</body>
</html>
